Move `ObjectUtils` to utils directory
Also use `plainObject` instead the old implementation.
